backend/requirements.txt
fastapi
starlette
typing
passlib
SQLAlchemy
python-jose
datetime